Initial commit for our new library datastock
Datastock is meant for automating data management within data science
and machine learning workflows. For this it provides a simple API that
allows to store and load data without having to care about where the
data will be saved and versioning it. It keeps the individual data
versions for each run and tracks the lineage of data to provide insights
about the dependencies within the workflow.

This commit contains a first working version which stores data in the
filesystem. Documentation will be added in a later commit, together with
additional features like providing some user interface for comparing
different runs and storing data in the cloud by using S3.