Complemented interface
This commit adds some missing functionality in the current interface. It
introduces a new function for requesting the info about a specific data
item. Additionally, we add a method to get the info from a data
reference and similar load the content from an info. To make the
concepts more clear, we rename DataItem to DataInfo to communicate that
it doesn't contain the data itself, but just information about it.
To resolve soem issues about circular imports, we move the stock
registry out to its own module.