Rename library to boxs
In order to keep the names in the style of real things in a workshop, we
rename it to boxs (without e, since 'boxes' is taken in PyPI). At the
same time, we replace the name of the Stock class, with Box, which makes
it shorter and additionally aligns well with the overall theme.