Create ValueType for represent different types of values
This commit introduces a new class ValueType that is used for
implementing the different ways to read or write values. Before we had
different functions for reading and writing, but no connection between
them if they were used for the same type. With the new class this
interface becomes more clear and we are able to choose the reading
method automatically, if we know how a value was written, to recreate
that value. This new class replaces the former inputs and outputs.