Added load() method to DataRef
We introduce a new method load() that allows to load the data directly
from a DataRef. Internally, we first fetch the info (if not available)
and then call load() on the DataInfo. This allows users to not have to
care about if they have a DataRef or DataInfo, the interface of the
possible methods is the same.