Add new methods to Storage interface to list items
In order to create a user interface for querying the stored data items
we need to create some functionality to inspect a storage and the items
in it. For this we create a new method to list all stored runs and a new
method for listing all data items that were created during a run.