Changed Storage to take DataRef instead of data_id and run_id
This commit changes the interface of Storage to make the methods take a
complete DataRef instead of data_id and run_ids. This simplifies the
interface and makes it possible in the same time for multiple boxes to
use the same storage. Support for this is currently still missing in the
FileSystemStorage, but this will be added at a later time.