Moved DataInfo creating into writer
This allows the writer to know about the individual attributes that end
up in the info, like tags or the name and to use them for additional
filtering or storing them in some kind of index. Before the writer had
to extract them from the info structure, which was brittle against
changes to the info format.