Add command line interface to inspect boxes
This commit adds a new command line interface "boxs" which can be used
for inspecting an existing box. It currently allows to list runs, list
the data items that were created by a specific run, print the
information about a data item as well as set names for runs, which
improves the usability.
The tool requires to be able to import the definition of the box from an
existing Python module.