Add new configuration class
This commit adds a new Configuration class that is automatically created
when imported from its module. The configuration allows to define a
default box, that will be used if no other box is specified, as well as
an init_module, that is automatically imported when defined. This module
mechanism allows to make sure, that all boxes are initialized before
they are being used. Both of these values can be configured using
environment variables using BOXS_DEFAULT_BOX and BOXS_INIT_MODULE. They
don't have to be set if they are not used.