Add new method to Storage for querying data items
This  commit introduces a new method `list_items(item_query)` to the
Storage interface, that takes a single ItemQuery as argument and allows
to query data items based on criteria like data id or name and run id
or name. We implement this method for the FileSystemStorage and add a
couple of tests for it.