Support as_file() with checksum
If we use a FileSystemStorage, we don't want every command inspecting
the storage copying over the file to a new location. Instead, if a local
file path is available we return the path within the storage directly.
This means, we can't create the checksum since we don't read the file.
So in the checksum transformer we ignore this special case under the
assumption, that a once written file doesn't change.