JsonValueType now supports lists and dicts
Previously the JsonValueType didn't support any automatic writing and
reading which made it only usable by explicitly defining it as argument
to a box.store(value_type=) argument. This is in practice quite loud and
not necessary. Therefore, we now support dicts and lists, so that they
can automatically stored and loaded.