Load the init_module after the whole boxs package is loaded
Using the init_module and loading it directly leads to problems with
circular imports, if the init_module accesses classes from boxs, which
is its sole purpose. Since the .config module is loaded before other
boxs modules, we have to wait with importing the init_module, to make
sure everything has been loaded and can be used.