Add new command to create dependency graph in DOT format
This new command allows to write a dependency graph of arbitrary items
to a file or to stdout. The language used for representing the graph is
DOT, which can be visualized with multiple tools. The command includes
besides the items matching the item query every ancestor, too. Each
run has its own subgraph, to make it easy to figure out, from which run
each individual data item originates.