Implement value type for storing tensorflow keras models
We create a new TensorflowKerasModelValueType that allows to store and
load tensorflow keras models. Since tensorflow would introduce a huge
number of additional dependencies, we import the functions for saving
and loading the model dynamically, so that we don't have to add
tensorflow as a new dependency. This means, that a missing tensorflow
library will raise an error, only if an instance of our new value type
is constructed.