Remove home directory from config and initialize lazily
This commit removes the home directory setting from the config. Using it
prevented us from logging the process of creating the config.
Additionally, we now set the config explicitly to initialized as part of
the get_config(). This fixes an issue, when multiple configs were
created, because creating a config lead to importing a python module,
that called get_config(), too, before the config was actually assigned.
Now the import happens after the config has been set to be initialized,
guaranteeing that the config object has been created and assigned.