Fix: test_main_list_runs_with_limit() now stable
The unit test was unstable because in it some files were written and the
order of some expected result depends on the order o the file
modification time. If the files have the same time the order is random.
Now we leave a pause of 10 ms between writing these files to make the
test deterministic.