Add new command to delete single run
The new command delete-run allows to delete a single run. This is
potential dangerous command because the user might delete a branch with
data, that is referenced from other branches.
Additionally, we fix the stream handling in the diff command by directly
using stdout and stderr without any pipe in between.