Moved transformer related classes to own module
This commit moves the Transformer base class and the delegating reader
and writer that are only used for transformers out of the storage module
to separate it from the "real" storage related classes. Someone who
would want to implement their own Storage class, shouldn't need to read
through unrelated code.