Moved creation of DataInfo out to Box.store()
Before the storage implementation was responsible for creating the
DataInfo that is returned from the Box.store() method when it calls
writer.write_info(). This lead to several problems:
  1) each Storage implementation had to do the same thing when creating
     the DataInfo.
  2) Storages shouldn't know about DataRef or DataInfo at all, since
     they are part of the user interface.
  3) Storage objects would need to understand what is part of the info.

With moving this responsibility out to Box.store() all these problems
disappear. The interface becomes much easier to understand and
implement.